+++ src/app/cart/cart.module.ts
@@ -1,21 +1,23 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { CartComponent } from './components/cart.component';
 import { CartService } from './services/cart.service';
+import { CartListComponent, CartItemComponent } from './components';
 
 @NgModule({
   imports: [
     CommonModule
   ],
   declarations: [
-    CartComponent
+    CartListComponent,
+    CartItemComponent
   ],
   providers: [
     CartService
   ],
   exports: [
-    CartComponent
+    CartListComponent,
+    CartItemComponent
   ]
 })
 export class CartModule { }