+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -1,0 +1,31 @@
+import { Component, OnInit } from '@angular/core';
+import { CartService } from '../../services/cart.service';
+import { CartItemModel } from '../models/cart-item.model';
+
+@Component({
+  selector: 'app-cart-list',
+  templateUrl: './cart-list.component.html',
+  styleUrls: ['./cart-list.component.css']
+})
+export class CartListComponent implements OnInit {
+  cartItems: Promise<Array<CartItemModel>>;
+
+  constructor(private cartService: CartService) { }
+
+  ngOnInit() {
+    this.cartItems = this.cartService.getCart();
+  }
+
+  onDeleteItem(cartItem: CartItemModel) {
+    this.cartService.removeItem(cartItem);
+  }
+
+  emptyCart() {
+    // this.cartItems = this.cartService.emptyCart();
+  }
+
+  checkout(): void {
+
+  }
+
+}