+++ src/app/cart/components/models/cart-item.interface.ts
@@ -1,0 +1,6 @@
+export interface ICartItem {
+  id: number;
+  name: string;
+  price: number;
+  qty: number;
+}